refactor to apply common ENV VARS
