+++ .gitignore
@@ -136,3 +136,6 @@ dmypy.json
 
 # jetbrains
 .idea
+
+# CDK outputs
+cdk.out/
