+++ config.py
@@ -1,0 +1,3 @@
+
+COST_CENTER = "COST_CENTER"
+COST_CENTER_TAG_NAME = "CostCenter"
